+++ smarts/core/traffic_history_provider.py
@@ -17,7 +17,6 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-import logging
 import numpy as np
 import sqlite3
 from itertools import cycle
@@ -34,7 +33,6 @@ class TrafficHistoryProvider(Provider):
     def __init__(self):
         self._histories = None
         self._is_setup = False
-        self._log = logging.getLogger(self.__class__.__name__)
         self._map_location_offset = None
         self._replaced_vehicle_ids = set()
         self._last_step_vehicles = set()
@@ -87,23 +85,6 @@ class TrafficHistoryProvider(Provider):
         # Ignore other sim state
         pass
 
-    def _decode_vehicle_type(self, vehicle_type: int) -> str:
-        # Options from NGSIM and INTERACTION currently include:
-        #  1=motorcycle, 2=auto, 3=truck, 4=pedestrian/bicycle
-        if vehicle_type == 1:
-            return "motorcycle"
-        elif vehicle_type == 2:
-            return "passenger"
-        elif vehicle_type == 3:
-            return "truck"
-        elif vehicle_type == 4:
-            return "pedestrian"
-        else:
-            self._log.warning(
-                f"unsupported vehicle_type ({vehicle_type}) in history data."
-            )
-        return "passenger"
-
     def step(
         self, provider_actions, dt: float, elapsed_sim_time: float
     ) -> ProviderState:
@@ -120,7 +101,7 @@ class TrafficHistoryProvider(Provider):
             if v_id in vehicle_ids or v_id in self._replaced_vehicle_ids:
                 continue
             vehicle_ids.add(v_id)
-            vehicle_type = self._decode_vehicle_type(hr.vehicle_type)
+            vehicle_type = self._histories.decode_vehicle_type(hr.vehicle_type)
             default_dims = VEHICLE_CONFIGS[vehicle_type].dimensions
             pos_x = hr.position_x + self._map_location_offset[0]
             pos_y = hr.position_y + self._map_location_offset[1]