+++ smarts/core/vehicle.py
@@ -277,15 +277,15 @@ class Vehicle:
         self._assert_initialized()
         # Assuming the position is the centre,
         # calculate the corner coordinates of the bounding_box
-        half_length = self.length / 2
-        half_width = self.width / 2
-        corners = [(-1, 1), (1, 1), (1, -1), (-1, -1)]
+        origin = self.position[:2]
+        dimensions = numpy.array([self.width, self.length])
+        corners = numpy.array([(-1, 1), (1, 1), (1, -1), (-1, -1)]) / 2
+        heading = self.heading
         return [
             rotate_around_point(
-                point=self.position[:2]
-                + numpy.array(corner) * numpy.array([half_width, half_length]),
-                radians=self.heading,
-                origin=self.position[:2],
+                point=origin + corner * dimensions,
+                radians=heading,
+                origin=origin,
             )
             for corner in corners
         ]
@@ -307,15 +307,23 @@ class Vehicle:
         controller_filepath,
         initial_speed=None,
     ):
-        # Agents can currently only control passenger vehicles
-        vehicle_type = "passenger"
-        chassis_dims = VEHICLE_CONFIGS[vehicle_type].dimensions
+        mission = mission_planner.mission
 
-        if isinstance(mission_planner.mission.task, UTurn):
-            if mission_planner.mission.task.initial_speed:
-                initial_speed = mission_planner.mission.task.initial_speed
+        if mission.vehicle_spec:
+            # mission.vehicle_spec.veh_type will always be "passenger" for now,
+            # but we use that value here in case we ever expand our history functionality.
+            vehicle_type = mission.vehicle_spec.veh_type
+            chassis_dims = mission.vehicle_spec.dimensions
+        else:
+            # non-history agents can currently only control passenger vehicles.
+            vehicle_type = "passenger"
+            chassis_dims = VEHICLE_CONFIGS[vehicle_type].dimensions
 
-        start = mission_planner.mission.start
+        if isinstance(mission.task, UTurn):
+            if mission.task.initial_speed:
+                initial_speed = mission.task.initial_speed
+
+        start = mission.start
         start_pose = Pose.from_front_bumper(
             front_bumper_position=numpy.array(start.position),
             heading=start.heading,
@@ -349,7 +357,11 @@ class Vehicle:
 
         chassis = None
         # change this to dynamic_action_spaces later when pr merged
-        if agent_interface and agent_interface.action in sim.dynamic_action_spaces:
+        if (
+            agent_interface
+            and agent_interface.action in sim.dynamic_action_spaces
+            and not mission.vehicle_spec
+        ):
             chassis = AckermannChassis(
                 pose=start_pose,
                 bullet_client=sim.bc,