+++ CHANGELOG.md
@@ -16,8 +16,11 @@ Copy and pasting the git commit messages is __NOT__ enough.
 ### Changed
 - `history_vehicles_replacement_for_imitation_learning.py` now uses new Imitation action space.  See Issue #844.
 - Updated and removed some package versions to ensure that Python3.8 is supported by SMARTS. See issue #266.
+=======
+- Refactored `Waypoints` into `LanePoints` (static, map-based) and `Waypoints` (dynamic).  See Issue #829.
 - Vehicles with a `BoxChassis` can now use an `AccelerometerSensor` too.
 - When importing NGSIM history data, vehicle speeds are recomputed.
+- Allow custom sizes for agent vehicles in history traffic missions.
 ### Fixed
 - Allow for non-dynamic action spaces to have action controllers.  See PR #854.
 - Fixed a minor bug in `sensors.py` which triggered `wrong_way` event when the vehicle goes into an intersection. See Issue #846.